+++ src/zkvyper/arguments.rs
@@ -53,7 +53,7 @@ pub struct Arguments {
     pub disable_vyper_optimizer: bool,
 
     /// Specify the path to the `vyper` executable. By default, the one in `${PATH}` is used.
-    /// In LLVM IR mode `vyper` is unused.
+    /// In LLVM IR and EraVM assembly modes, `vyper` executable is unused.
     #[structopt(long = "vyper")]
     pub vyper: Option<String>,
 
@@ -67,12 +67,12 @@ pub struct Arguments {
     #[structopt(long = "enable-decimals")]
     pub enable_decimals: bool,
 
-    /// An extra output format string.
-    /// See `vyper --help` for available options including combined JSON mode.
+    /// Set the output format selection.
+    /// Available options: combined_json | ir_json | metadata | ast | abi | method_identifiers | layout | userdoc | devdoc | eravm_assembly
     #[structopt(short = "f")]
     pub format: Option<String>,
 
-    /// Sets the number of threads, which execute the tests concurrently.
+    /// Set the number of threads, which execute the tests concurrently.
     #[structopt(short = "t", long = "threads")]
     pub threads: Option<usize>,